Fix server failing to generate games on Windows
The old standard way of providing simulate.py with the paths to the Gemini core ASP files
(by expanding ./common.sh on the command line every time we invoke simulate.py)
doesn't work on Windows without a Unix-like shell installed. Instead, we can use
a single ASP file that imports all the others (gemini.lp) as a single entry point
to the ASP portion of the generator.